+++ src/Widget/RadioList/RadioList.php
@@ -256,7 +256,7 @@ final class RadioList implements NoEncodeStringableInterface
                 Html::getNonArrayableName($this->name),
                 $this->uncheckValue
             )
-                ->attributes(
+                ->addAttributes(
                     array_merge(
                         [
                             // Make sure disabled input is not sending any value
@@ -275,8 +275,7 @@ final class RadioList implements NoEncodeStringableInterface
             return ($this->itemFormatter)($item);
         }
 
-        $radio = Html::radio($item->name, $item->value)
-            ->attributes($item->radioAttributes)
+        $radio = Html::radio($item->name, $item->value, $item->radioAttributes)
             ->checked($item->checked)
             ->label($item->label)
             ->labelEncode($item->encodeLabel);